requirements.txt
pynput
psutil
Pillow
pywin32
